extracted workspace directory logic from index
- moved workspace directory logic from the validate function in index
  to a new function named executeInWorkspaceDirectory in util
